Separate out Critic and Replay Buffer
